Refatorado alguns códigos
