--- src/components/TweetScreen.tsx
@@ -1,1 +1,0 @@
-export const TweetScreen = () => {};
