+++ src/store/store-context.tsx
@@ -1,10 +1,13 @@
 import React, { useContext, createContext } from "react";
+import { Tweet } from "../components/Home";
 
 export interface DarkContext {
   isDark: boolean;
   isLogedIn: boolean;
   setIsDark: () => void;
   onChangeLogin: () => void;
+  arrayTweets: Tweet[];
+  onLocalStorage: () => void;
 }
 
 export const darkContext = createContext<DarkContext>({
@@ -12,4 +15,6 @@ export const darkContext = createContext<DarkContext>({
   setIsDark: () => {},
   onChangeLogin: () => {},
   isLogedIn: false,
+  arrayTweets: [],
+  onLocalStorage: () => {},
 });
